Release finale

+++ server.js
@@ -51,7 +51,7 @@ function useTemplate(res,path,options){
 }
 
 //filtro che si occupa reindirizzare alla home tutte le richieste "generiche" del tipo /
-app.use('/$',function(req,res){
+app.get('/$',function(req,res){
 	//reindirizzo al middleware che si occupa della home page
 	res.redirect('/home');
 });
@@ -65,39 +65,52 @@ app.use('/private/*', function(req,res,next){
 });
 
 //ritorno pagina di login con il messaggio standard di richiesta delle credenziali
-app.use('/login', function(req,res){
+app.get('/login', function(req,res){
 	//messaggioErrore è il messaggio che verrà mostrato sopra il form 
 	useTemplate(res,'private/login.html',{messaggioErrore: "Inserisci le tue credenziali"});
 });
 
 //funzione di autenticazione
 app.post('/accedi', function(req,res){
-	
-	//dichiaro la query che dovrà essere eseguita con gli specificatori di formato $1 e $2,
-	//che saranno automaticamente rimpiazzati dai valori che l'utente ha inserito nel form
-	var query = { text: 'select * from utenti where username = $1 and password = $2',
-			//grazie a body-parser, accedere ai campi del form è davvero semplice: req.body.nome_dell_input
-			values: [req.body.username,req.body.password] 
-	};
-	
-	//lancio la funzione launchQuery definita in db.js che permette di eseguire una query ed infine chiama la funzione come secondo parametro
-	//result è l'array che contiene le righe risultate dalla query
-	db.launchQuery(query, function(err,result){
-		//se c'è stato un errore, rimando a error
-		if(err) res.redirect('/error');
-		//altrimenti controllo che sia stato trovato un utente con quel username e quella password
-		else if(result.length==1){
-			req.session.user = new obj.utente(result[0].id,result[0].username,result[0].via);
-			res.redirect('/home');
-		}
-		//infine, se utente e password erano errati, rimando al login con un messaggio 
-		else {
+	console.log("Entrooo");
+	console.log(req.body.username + ' ' + req.body.password);
+	if(req.body != undefined){
+		if(req.body.username != undefined && req.body.username != '' && req.body.password != undefined && req.body.password != ''){
+			console.log('Request: ' + req.body.username + ',' + req.body.password);
+			//dichiaro la query che dovrà essere eseguita con gli specificatori di formato $1 e $2,
+			//che saranno automaticamente rimpiazzati dai valori che l'utente ha inserito nel form
+			var query = { text: 'select * from utenti where username = $1 and password = $2',
+					//grazie a body-parser, accedere ai campi del form è davvero semplice: req.body.nome_dell_input
+					values: [req.body.username,req.body.password] 
+			};
+			
+			//lancio la funzione launchQuery definita in db.js che permette di eseguire una query ed infine chiama la funzione come secondo parametro
+			//result è l'array che contiene le righe risultate dalla query
+			db.launchQuery(query, function(err,result){
+				//se c'è stato un errore, rimando a error
+				if(err) res.redirect('/error');
+				//altrimenti controllo che sia stato trovato un utente con quel username e quella password
+				else if(result.length==1){
+					req.session.user = new obj.utente(result[0].id,result[0].username,result[0].via);
+					res.redirect('/home');
+				}
+				//infine, se utente e password erano errati, rimando al login con un messaggio 
+				else {
+					useTemplate(res,'private/login.html',{
+						//tx.menu ritorna il codice html della tabella generata per mostrare i dati di result
+						messaggioErrore: "Username o password errati"
+					});
+				}
+			});
+		} else{
 			useTemplate(res,'private/login.html',{
 				//tx.menu ritorna il codice html della tabella generata per mostrare i dati di result
 				messaggioErrore: "Username o password errati"
 			});
 		}
-	});
+	} else {
+		res.redirect('/error');
+	}
 });
 
 //ritono la home page modificata con i parametri dell'utente loggato in modo che sia personalizzabile
@@ -205,19 +218,21 @@ app.get('/private/choose', function(req,res){
 
 //inserisce nel database le scelte fatte dall'utente nella scelta del menu per il giorno successivo
 app.post('/private/choose',function(req,res){
-	//creo un array di query che dovranno essere eseguite dalla funzione launchDeepQuey
-	var query = [];
-	for(c in req.body){
-		query.push({
-			text: 'insert into sceglie (id_utente,id_menu,data) values ($1,$2,now() + interval \'1 day\')',
-			values: [req.session.user.id,parseInt(req.body[c])]
+	if(req.body != undefined){
+		//creo un array di query che dovranno essere eseguite dalla funzione launchDeepQuey
+		var query = [];
+		for(c in req.body){
+			query.push({
+				text: 'insert into sceglie (id_utente,id_menu,data) values ($1,$2,now() + interval \'1 day\')',
+				values: [req.session.user.id,parseInt(req.body[c])]
+			});
+		}
+		//lancio le query dicendo di partire dalla prima (0) nell'array e di eseguire, una volta terminato, la funzione come secondo paramentro
+		db.launchDeepQuery(query,0,function(err) {
+			if (err) res.redirect('/error');		//in caso di errore rimando a /error
+			else res.redirect('/home');				//in caso di successo rimando alla home
 		});
-	}
-	//lancio le query dicendo di partire dalla prima (0) nell'array e di eseguire, una volta terminato, la funzione come secondo paramentro
-	db.launchDeepQuery(query,0,function(err) {
-		if (err) res.redirect('/error');		//in caso di errore rimando a /error
-		else res.redirect('/home');				//in caso di successo rimando alla home
-	});
+	} else res.redirect('/error');
 });
 
 //middleware per rimuovere dal database tutte le prenotazioni fatte per l'indomani, quando l'utente dedice di volerle rifare
@@ -250,43 +265,52 @@ app.get('/private/vota',function(req,res){
 //middleware che si occupa della registrazione dei voti sul database
 //i voti vengono controllati perchè quelli con valore 0 non hanno senso ad essere riscritti uguali nel db
 app.post('/private/addVoto',function(req,res){
-	//creo array di query da eseguire per aggiornare i voti sul db
-	var query = [];
-	for(c in req.body){
-		if(parseInt(req.body[c])!=0){
-			query.push({
-				text: 'update sceglie set voto = $1 where id_utente = $2 and id_menu = $3',
-				values: [parseInt(req.body[c]),req.session.user.id,parseInt(c)]
-			});
+	if(req.body != undefined){
+		//creo array di query da eseguire per aggiornare i voti sul db
+		var query = [];
+		for(c in req.body){
+			if(parseInt(req.body[c])!=0){
+				query.push({
+					text: 'update sceglie set voto = $1 where id_utente = $2 and id_menu = $3',
+					values: [parseInt(req.body[c]),req.session.user.id,parseInt(c)]
+				});
+			}
 		}
-	}
-	//lancio la serie di query con la funzione launchDeepQuery, 
-	//che al termine di tutto il suo lavoro chiama la funzione passata come secondo argomento
-	db.launchDeepQuery(query,0,function(err) {
-		if (err) res.redirect('/error');
-		else res.redirect('/private/vota');
-	});
+		//lancio la serie di query con la funzione launchDeepQuery, 
+		//che al termine di tutto il suo lavoro chiama la funzione passata come secondo argomento
+		db.launchDeepQuery(query,0,function(err) {
+			if (err) res.redirect('/error');
+			else res.redirect('/private/vota');
+		});
+	} else res.redirect('/error');
 });
 
 //middleware chiamato per aggiungere un'allergia dell'utente corrente sul database
 app.post('/private/addAllergia', function(req,res){
-	db.launchQuery({
-		text: 'insert into intollerante (id_utente,id_allergie) values ($1,$2);',
-		values: [req.session.user.id,parseInt(req.body.id_allergie)]
-	}, function(err, result) {
-		res.redirect('/private/allergie');
-	});
+	console.log('Adding');
+	if(req.body != undefined){
+		if((req.body.id_allergie != undefined) && (typeof req.body.id_allergie === 'string')){
+			db.launchQuery({
+				text: 'insert into intollerante (id_utente,id_allergie) values ($1,$2);',
+				values: [req.session.user.id,parseInt(req.body.id_allergie)]
+			}, function(err, result) {
+				res.redirect('/private/allergia');
+			});
+		} else res.redirect('/private/allergia');
+	} else res.redirect('/error');
 });
 
 //middleware chiamato per rimuovere un'allergia dell'utente corrente sul database
 app.post('/private/removeAllergia', function(req,res){
-	db.launchQuery({
-		text: 'delete from intollerante where id_utente = $1 and id_allergie = $2;',
-		values: [req.session.user.id,parseInt(req.body.id_allergie)]
-	}, function(err, result) {
-		if (err) res.redirect('/error');
-		else res.redirect('/private/allergie');
-	});
+	if(req.body != undefined){
+		db.launchQuery({
+			text: 'delete from intollerante where id_utente = $1 and id_allergie = $2;',
+			values: [req.session.user.id,parseInt(req.body.id_allergie)]
+		}, function(err, result) {
+			if (err) res.redirect('/error');
+			else res.redirect('/private/allergie');
+		});
+	} else res.redirect('/error');
 });
 
 app.get('/private/allergie', function(req,res){
@@ -305,7 +329,8 @@ app.get('/private/allergie', function(req,res){
 				else {
 					useTemplate(res,'private/gestisciAllergie.html',
 					{
-						tabella: tx.menuAllergie(result),	//ritorna una semplice tabella HTML per mostrare le allergie
+						//ritorna una semplice tabella HTML per mostrare le allergie
+						tabella: tx.menuAllergie(result),	
 						//aggiungo un select-field per poter scegliere ed aggiungere una nuova allergia all'utente corrente
 						//ed un pulsante per confermare la scelta
 						selezione: tx.buttonAllergie(allergie)
@@ -319,12 +344,16 @@ app.get('/private/allergie', function(req,res){
 //middleware chiamato nel caso in cui un errore sia stato incontrato dagli altri middleware o dal database nell'elaborazione delle richieste
 //la dichiarazione poteva essere inserita come statica in express ma in implementazioni future potrebbe 
 //trasformarsi in un template per mostrare lo specifico errore anche all'utente
-app.use('/error',function(req,res){
-	res.sendFile('error_page.html',{ root : __dirname});
+app.get('/error',function(req,res){
+	//ritorno la pagina di errore e codice 406
+	res.writeHead(406, {'Content-Type':'text/html'});
+	bind.toFile('error_page.html',{},function(data){
+		res.end(data);
+	});
 });
 
 //metto in ascolto il sistema sulla porta precedentemente settata ed eseguo la funzione anonima per mostrare come accedere al sistema tramite browser
 app.listen(app.get('port'), function() {
-  console.log('Node app is running on port', app.get('port'));
+  	console.log('Node app is running on port', app.get('port'));
 });
